Did heap Coderbyte 2 but without explanations or full understanding yet
I’m still confused about max and min and the default heapq. I also need to verify I understand the underlying list when doing heaps and their functions.

I believe I understand it because of the Coachable lectures and reading, but yeah.

So like remembering ascending and descending. The default heapq heapify and heappush are always putting the smallest first/at the top/where [0] or lower index is. This is also what gets popped first, right? So by default it is ascending?